refactor 클래스 분리 및 수정
 - View 패키지를 새로 만들고 그 아래에 Input/Output 처리하는 클래스
   두 개로 분리
 - Check를 ElementValidator로 변경함
 - equationDelimiter을 상수로 변경함
