+++ src/main/java/calculator/view/OutputView.java
@@ -1,0 +1,7 @@
+package calculator.view;
+
+public class OutputView {
+    public void showResult(double result) {
+        System.out.println(result);
+    }
+}
